refactor(live-activity): extract checkLiveActivity probe into @wxyc/database (closes #1241)
Three callers (warm-walker boot pass, flowsheet-metadata-backfill cron,
flowsheet-artwork-repair drain) had byte-equivalent copies of the
cooperative-pause probe — same SELECT, same schema-sanitization helper,
same 60s/30s defaults. Move the probe + defaults + CheckLiveActivityFn
type to shared/database/src/live-activity.ts and have each caller import
from @wxyc/database.

Per-caller env resolvers stay in place: warm.service reads
WARM_LIVE_ACTIVITY_LOOKBACK_SECONDS and warns-on-invalid (API boot must
succeed), while the two job orchestrators read
LIVE_ACTIVITY_LOOKBACK_SECONDS and throw-on-invalid (cron should exit
loud on misconfig). Throw-vs-warn is a real policy difference, not
duplication.

album-level-backfill is deliberately left out: its inline comment
documents the divergence (single 3-hour transaction lifecycle, 300s
lookback override) and its awaitQuietWindow loop shape differs from the
other three's inline while-await pattern. Extracting it would couple
two intentionally-different envelopes.

Net -72 lines across the four touched files.
