+++ shared/database/src/index.ts
@@ -5,3 +5,4 @@ export * from './legacy/sql.mirror.js';
 export * from './legacy/etl-utils.js';
 export * from './cdc-listener.js';
 export * from './library-tiebreak.js';
+export * from './live-activity.js';
