+++ tests/mocks/database.mock.ts
@@ -267,6 +267,14 @@ export const flowsheetEntryTypeEnum = () => ({});
 // which library_id the tie-break "picks".
 export const pickPrimaryLibraryRow = jest.fn<(libraryIds: number[]) => Promise<number | null>>();
 
+// Cooperative-pause primitives (BS#1241). Mirrored from shared/database/src/live-activity.ts
+// so consumer unit tests that import them via `@wxyc/database` get the real default values
+// rather than `undefined`.
+export const LIVE_ACTIVITY_LOOKBACK_SECONDS_DEFAULT = 60;
+export const LIVE_ACTIVITY_PAUSE_MS_DEFAULT = 30_000;
+export type CheckLiveActivityFn = (lookbackSeconds: number) => Promise<boolean>;
+export const checkLiveActivity = jest.fn<CheckLiveActivityFn>().mockResolvedValue(false);
+
 // Mock types
 export type AnonymousDevice = {
   id: number;
